# www_app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for the Laravel application.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pet-breeds": {
            "get": {
                "tags": [
                    "PetBreeds"
                ],
                "summary": "Get list of Pet Breeds",
                "operationId": "718fea5037700adc791569d16e8476e3",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "pet_type_id",
                        "in": "query",
                        "description": "ID of the Pet Breed to filter breeds",
                        "required": true,
                        "schema": {
                            "description": "Must be a valid ID from the pet_types table",
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Китайская хохлатая собака"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-03-12T20:08:04.566Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "PetBreeds"
                ],
                "summary": "Store a new Pet Breed",
                "description": "Creates a new Pet Breed and stores it in the database.",
                "operationId": "storePetBreeds",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pet_type_id",
                                    "name_uk",
                                    "name_en",
                                    "name_ru"
                                ],
                                "properties": {
                                    "pet_type_id": {
                                        "description": "ID of the Pet Breed",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name_uk": {
                                        "description": "Name of the created Pet Breed in Ukrainian",
                                        "type": "string",
                                        "example": "Китайський чубатий собака"
                                    },
                                    "name_en": {
                                        "description": "Name of the created Pet Breed in English",
                                        "type": "string",
                                        "example": "Chinese Crested Dog"
                                    },
                                    "name_ru": {
                                        "description": "Name of the created Pet Breed in Russian",
                                        "type": "string",
                                        "example": "Китайская хохлатая собака"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pet Breed created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID of the created Pet Breed",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "pet_type_id": {
                                            "description": "ID of the Pet Breed",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name_uk": {
                                            "description": "Name of the created Pet Breed in Ukrainian",
                                            "type": "string",
                                            "example": "Китайський чубатий собака"
                                        },
                                        "name_en": {
                                            "description": "Name of the created Pet Breed in English",
                                            "type": "string",
                                            "example": "Chinese Crested Dog"
                                        },
                                        "name_ru": {
                                            "description": "Name of the created Pet Breed in Russian",
                                            "type": "string",
                                            "example": "Китайская хохлатая собака"
                                        },
                                        "updated_at": {
                                            "description": "Date and time of the last update",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-12T20:08:04.566Z"
                                        },
                                        "created_at": {
                                            "description": "Date and time of the creation",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-12T20:08:04.566Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}